Add Renderpass object
Granite has this as a rendergraph, but that somewhat misleads a user as to the use of the object. It also makes it less flexible. 

A renderpass does some rendergraph-esque tasks to handle the correct setup of subpasses, but there will still be another rendergraph object taking care of higher-level and more abstract tasks.